refactor: 유저 관련 페이지 뒤로 가기 컴포넌트화
