+++ src/app/(user)/components/BackBtn.stories.tsx
@@ -2,24 +2,33 @@ import type { Meta, StoryObj } from '@storybook/react';
 import { BackBtn } from './BackBtn';
 
 const meta: Meta<typeof BackBtn> = {
-  title: 'Recently-viewed/BackBtn',
+  title: 'User/BackBtn',
   component: BackBtn,
   parameters: {
     layout: 'centered',
   },
   tags: ['autodocs'],
-
   decorators: [
     (Story) => (
-      <div className='w-32'>
+      <div className="flex w-32">
         <Story />
       </div>
     ),
   ],
+  argTypes: {
+    url: {
+      control: 'text',
+      description: '이동할 페이지의 URL',
+    },
+  },
 };
 
 export default meta;
 type Story = StoryObj<typeof BackBtn>;
 
 // 기본 스토리
-export const Default: Story = {};
+export const Default: Story = {
+  args: {
+    url: '/',
+  },
+};
