+++ src/app/(user)/wishlists/[id]/page.tsx
@@ -1,7 +1,8 @@
-import { BackBtn } from '@/app/(user)/wishlists/[id]/components/BackBtn';
+import { BackBtn } from '@/app/(user)/components/BackBtn';
 import { Btn } from '@/app/(user)/wishlists/[id]/components/Btn';
 import { SettingBtn } from '@/app/(user)/wishlists/[id]/components/SettingBtn';
 import { WishlistCard } from '@/app/(user)/wishlists/[id]/components/WishlistCard';
+import { ROUTES } from '@/constants/routeURL';
 
 const filters = ['날짜 입력하기', '게스트 1명', '공유하기'];
 
@@ -10,7 +11,7 @@ export default function WishlistsDetail() {
     <div className="flex w-full">
       <div className="w-[65%] px-5 pt-5">
         <div className="flex w-full items-center justify-between">
-          <BackBtn />
+          <BackBtn url={ROUTES.USER.WISHLISTS} />
           <SettingBtn />
         </div>
         <h1 className="my-5 text-2xl font-medium">위시리스트 폴더 명</h1>
